+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2026-03-23
+
+### Added
+
+- Add `shutdown` action for Android emulators and iOS simulators.
+
+### Changed
+
+- Change default keymap for `launch` and `launch with option` actions for Android emulators.
+
 ## [0.2.1] - 2026-03-21
 
 ### Fixed